backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=24.1.0",
    "alembic>=1.16.4",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "openpyxl>=3.1.5",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pymupdf>=1.26.3",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.42",
    "unstructured>=0.18.11",
    "uvicorn[standard]>=0.35.0",
    "asyncpg>=0.30.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "aiosqlite>=0.21.0",
]
